Create type HeaderSubtext
We'll use this for now to represent subtext with inlined headers.
